move layouts to it's own directory under src
