+++ src/layout/MainLayout.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Head from 'next/head';
-import NavigationPanel from '../../components/navigation-panel/NavigationPanel';
-import Footer from '../../components/footer/Footer';
+import NavigationPanel from '../components/navigation-panel/NavigationPanel';
+import Footer from '../components/footer/Footer';
 
 const MainLayout: React.FC = ({ children }) => {
   return (
